+++ packages/@ionic/cli-framework/src/lib/colors.ts
@@ -1,12 +1,10 @@
+import { ColorFunction, Colors as BaseColors } from '@ionic/cli-framework-output';
 import * as chalk from 'chalk';
 import * as lodash from 'lodash';
 
 import { MetadataGroup } from '../definitions';
 
-import { LoggerLevel } from './logger';
-
-export type ColorFunction = (...text: string[]) => string;
-export type LoggerColors = { [L in LoggerLevel]: ColorFunction; };
+export { ColorFunction, LoggerColors } from '@ionic/cli-framework-output';
 export type HelpGroupColors = { [G in Exclude<MetadataGroup, MetadataGroup.HIDDEN | MetadataGroup.ADVANCED>]: ColorFunction; };
 
 export interface HelpColors {
@@ -18,39 +16,7 @@ export interface HelpColors {
   group: HelpGroupColors;
 }
 
-export interface Colors {
-  /**
-   * Used to mark text as important. Comparable to HTML's <strong>.
-   */
-  strong: ColorFunction;
-
-  /**
-   * Used to mark text as less important.
-   */
-  weak: ColorFunction;
-
-  /**
-   * Used to mark text as input such as commands, inputs, options, etc.
-   */
-  input: ColorFunction;
-
-  /**
-   * Used to mark text as successful.
-   */
-  success: ColorFunction;
-
-  /**
-   * Used to mark text as failed.
-   */
-  failure: ColorFunction;
-
-  /**
-   * Used to mark text as ancillary or supportive.
-   */
-  ancillary: ColorFunction;
-
-  log: LoggerColors;
-
+export interface Colors extends BaseColors {
   help: HelpColors;
 }