+++ packages/@ionic/cli-framework/src/lib/index.ts
@@ -1,15 +1,8 @@
-import { Logger } from './logger';
-
 export * from './colors';
 export * from './command';
 export * from './completion';
 export * from './config';
 export * from './executor';
 export * from './help';
-export * from './logger';
 export * from './options';
-export * from './output';
-export * from './tasks';
 export * from './validators';
-
-export const logger = new Logger();