+++ packages/@ionic/cli-framework/src/utils/string.ts
@@ -19,10 +19,6 @@ export function isValidURL(url?: any): boolean {
   return URL_REGEX.test(url);
 }
 
-export function enforceLF(str: string): string {
-  return str.match(/[\r\n]$/) ? str : str + '\n';
-}
-
 export function strcmp(a: string | undefined, b: string | undefined): number {
   if (!a) {
     a = '';