+++ packages/@ionic/cli/src/commands/deploy/build.ts
@@ -1,4 +1,5 @@
-import { CommandLineInputs, CommandLineOptions, LOGGER_LEVELS, MetadataGroup } from '@ionic/cli-framework';
+import { CommandLineInputs, CommandLineOptions, MetadataGroup } from '@ionic/cli-framework';
+import { LOGGER_LEVELS } from '@ionic/cli-framework-output';
 import { columnar } from '@ionic/cli-framework/utils/format';
 import { sleep } from '@ionic/utils-process';
 import * as chalk from 'chalk';
@@ -36,6 +37,7 @@ export class BuildCommand extends Command {
     return {
       name: 'build',
       type: 'project',
+      groups: [MetadataGroup.PAID],
       summary: 'Create a deploy build on Appflow',
       description: `
 This command creates a deploy build on Ionic Appflow. While the build is running, it prints the remote build log to the terminal.
@@ -158,17 +160,29 @@ Apart from ${input('--commit')}, every option can be specified using the full na
     const ws = this.env.log.createWriteStream(LOGGER_LEVELS.INFO, false);
 
     let isCreatedMessage = false;
+    let errorsEncountered = 0;
     while (!(build && (build.state === 'success' || build.state === 'failed'))) {
-      await sleep(5000);
-      build = await this.getDeployBuild(appflowId, buildId, token);
-      if (build && build.state === 'created' && !isCreatedMessage) {
-        ws.write(chalk.yellow('Concurrency limit reached: build will start as soon as other builds finish.'));
-        isCreatedMessage = true;
-      }
-      const trace = build.job.trace;
-      if (trace.length > start) {
-        ws.write(trace.substring(start));
-        start = trace.length;
+      try {
+        await sleep(5000);
+        build = await this.getDeployBuild(appflowId, buildId, token);
+        if (build && build.state === 'created' && !isCreatedMessage) {
+          ws.write(chalk.yellow('Concurrency limit reached: build will start as soon as other builds finish.'));
+          isCreatedMessage = true;
+        }
+        const trace = build.job.trace;
+        if (trace.length > start) {
+          ws.write(trace.substring(start));
+          start = trace.length;
+        }
+        errorsEncountered = 0;
+      } catch (e) {
+        // Retry up to 3 times in the case of an error.
+        errorsEncountered++;
+        ws.write(chalk.yellow(`Encountered error: ${e} while fetching build data retrying.`));
+        if (errorsEncountered >= 3) {
+          ws.write(chalk.red(`Encountered ${errorsEncountered} errors in a row. Job will now fail.`));
+          throw e;
+        }
       }
     }
     ws.end();