+++ packages/@ionic/cli-framework/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.3](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@5.0.2...@ionic/cli-framework@5.0.3) (2020-08-29)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
+## [5.0.2](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@5.0.1...@ionic/cli-framework@5.0.2) (2020-08-28)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
+## [5.0.1](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@5.0.0...@ionic/cli-framework@5.0.1) (2020-08-27)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
+# [5.0.0](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.2.2...@ionic/cli-framework@5.0.0) (2020-08-27)
+
+
+### Code Refactoring
+
+* do not re-export from @ionic/cli-framework-output ([a91b5a4](https://github.com/ionic-team/ionic-cli/commit/a91b5a4cb76570154e560bdea3138a425833ce8c))
+
+
+### BREAKING CHANGES
+
+* Install `@ionic/cli-framework-output` and import from it directly.
+
+
+
+
+
+## [4.2.2](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.2.1...@ionic/cli-framework@4.2.2) (2020-08-26)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
+## [4.2.1](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.2.0...@ionic/cli-framework@4.2.1) (2020-08-25)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
 # [4.2.0](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.1.5...@ionic/cli-framework@4.2.0) (2020-06-02)