+++ packages/@ionic/utils-object/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.5](https://github.com/ionic-team/ionic-cli/compare/@ionic/utils-object@2.1.4...@ionic/utils-object@2.1.5) (2020-08-28)
+
+**Note:** Version bump only for package @ionic/utils-object
+
+
+
+
+
+## [2.1.4](https://github.com/ionic-team/ionic-cli/compare/@ionic/utils-object@2.1.3...@ionic/utils-object@2.1.4) (2020-08-25)
+
+**Note:** Version bump only for package @ionic/utils-object
+
+
+
+
+
 ## [2.1.3](https://github.com/ionic-team/ionic-cli/compare/@ionic/utils-object@2.1.2...@ionic/utils-object@2.1.3) (2020-05-12)